ai/requirements.txt
opencv-python
mediapipe
pyyaml
numpy

joblib

tensorflow